% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_border.R
\name{add_border}
\alias{add_border}
\alias{add_header}
\alias{add_footer}
\title{Add Border to Slides}
\usage{
add_border(inverse = FALSE)

add_header(inverse = FALSE)

add_footer(inverse = FALSE)
}
\arguments{
\item{inverse}{Default FALSE; should we use the inverse theme?}
}
\description{
Functions to add a header, footer or both.
}
