% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_class.R
\name{add_class}
\alias{add_class}
\title{Add class tags to the slides}
\usage{
add_class(inverse = FALSE, center = FALSE, middle = FALSE)
}
\arguments{
\item{inverse}{Default \code{FALSE}}

\item{center}{Default \code{FALSE}}

\item{middle}{Default \code{FALSE}}
}
\description{
Add class tags to the slides
}
