% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_slide.R
\name{build_slide}
\alias{build_slide}
\title{Convert RMD to HTML}
\usage{
build_slide(fname)
}
\arguments{
\item{fname}{Filename}
}
\description{
If the slides are not runtime shiny, the filename is passed to
 \code{rmarkdown::render}.
}
\details{
This could be run in parallel, but currently isn't
}
