% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{create_connect_template}
\alias{create_connect_template}
\title{Copy the index.Rmd template for the connect back up to local directory}
\usage{
create_connect_template()
}
\description{
Should only be used in a Makefile in slides directory
}
