% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_example.R
\name{create_example}
\alias{create_example}
\title{Create Example Presentation}
\usage{
create_example(path = NULL)
}
\arguments{
\item{path}{A path where the presentation should be copied to.}
}
\description{
Copy example presentation to a directory specified by \code{path}
}
