% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_feedback.R
\name{create_feedback}
\alias{create_feedback}
\title{Create a feedback slide}
\usage{
create_feedback()
}
\description{
Copies the feedback slide template to the current WD. This feedback slide will look for the config questionnaire item
}
