% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{upload_slides}
\alias{upload_slides}
\alias{create_site_yml}
\alias{clean_site}
\title{Create site yaml for connect}
\usage{
upload_slides(clean = TRUE, user = Sys.getenv("JR_CONNECT_USER"))

create_site_yml()

clean_site()
}
\arguments{
\item{clean}{Should we clean up?}

\item{user}{Connect user. Default \code{Sys.getenv("JR_CONNECT_USER")}}
}
\description{
Should only be used in a Makefile in slides directory
}
