% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_welcome.R
\name{create_welcome}
\alias{create_welcome}
\title{Create Example Presentation}
\usage{
create_welcome(path = ".")
}
\arguments{
\item{path}{A path where the presentation should be copied to.}
}
\description{
Copy example presentation to a directory specified by \code{path}
}
