% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_author.R
\name{get_author}
\alias{get_author}
\title{Presentation author and twitter handle}
\usage{
get_author(user = NULL)
}
\arguments{
\item{user}{By default grabs the system name and matches user.}
}
\description{
Returns a nicely formatted name and twitter handle.
}
\examples{
get_author(user = "ncsg3")
}
