% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_user}
\alias{get_user}
\title{Presentation user_name for rstudio connect}
\usage{
get_user(user = NULL)
}
\arguments{
\item{user}{By default grabs the system name and matches user.}
}
\description{
Returns a nicely formatted name and twitter handle.
}
\examples{
get_user(user = "ncsg3")
}
