% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_options.R
\name{set_presentation_options}
\alias{set_presentation_options}
\title{Set knitr Options}
\usage{
set_presentation_options(page_numbers = FALSE, url = "")
}
\arguments{
\item{page_numbers}{Boolean Add page numbers to the slides}

\item{url}{Character (default "") optional link to package}
}
\description{
In the future, pass the options as arguments.
}
