% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setnicepar.R
\name{setnicepar}
\alias{setnicepar}
\title{Set a nice par}
\usage{
setnicepar(mar = c(3, 3, 2, 1), mgp = c(2, 0.4, 0), tck = -0.01,
  cex_axis = 0.9, las = 1, mfrow = c(1, 1), ...)
}
\arguments{
\item{mar}{See \code{?par}}

\item{mgp}{See \code{?par}}

\item{tck}{See \code{?par}}

\item{cex_axis}{See \code{?par}}

\item{las}{See \code{?par}}

\item{mfrow}{See \code{?par}}

\item{...}{Additional arguments passed to \code{par}}
}
\description{
Colin's wonderful setnicepar functions for sexy base graphics
}
\examples{
plot(1:10)
setnicepar()
plot(1:10)
}
