% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lint_slides}
\alias{lint_slides}
\title{Lintr functions}
\usage{
lint_slides(fnames = NULL)
}
\arguments{
\item{fnames}{An optional vector of filenames}
}
\description{
Run lintr on a vector of files. If \code{fnames} is \code{NULL}
then use the regular expression \code{^c.*Rmd$}. If
\code{config::get("lintr")} is \code{FALSE}, no lint used.
}
